+++ src/features/login/ui/LoginForm.tsx
@@ -1,13 +1,13 @@
 import { Checkbox, Flex, Text } from '@chakra-ui/react';
 import { useRouter } from '@tanstack/react-router';
 import { useLoginMutation } from 'features/login/api/login.mutation';
-import { LoginFormData, LoginStorageModel } from 'features/login/model/type';
-import { useEffect, useState } from 'react';
+import { LoginFormData } from 'features/login/model/type';
 import { SubmitHandler } from 'react-hook-form';
 import { Button, Input } from 'shared/ui/FormControl';
 
 import { LOGIN_INPUT_LIST } from '../lib/loginInputList';
 import { useLoginForm } from '../lib/useLoginForm';
+import { loginStorageModel } from '../model/LoginStorageModel';
 
 const LoginForm = () => {
   const {
@@ -15,8 +15,6 @@ const LoginForm = () => {
     state: { location },
     history,
   } = useRouter();
-
-  const [storage, setStorage] = useState<LoginStorageModel | null>(null);
 
   const { registerField, errors, setError, handleSubmit, isValid } =
     useLoginForm();
@@ -35,8 +33,8 @@ const LoginForm = () => {
       { loginRequest: { ...data } },
       {
         onSuccess: (res) => {
-          storage?.setEmail(data.email);
-          storage?.setToken(res.token);
+          loginStorageModel?.setEmail(data.email);
+          loginStorageModel?.setToken(res.token);
           navigateLocation();
         },
         onError: (error) => {
@@ -51,11 +49,6 @@ const LoginForm = () => {
       },
     );
   };
-
-  useEffect(() => {
-    const storage = new LoginStorageModel();
-    setStorage(storage);
-  }, []);
 
   return (
     <form onSubmit={handleSubmit(onLogin)}>
